+++ NEWS.md
@@ -1,4 +1,8 @@
-# teal 0.11.1.9023
+# teal 0.11.1.9024
+
+### New features
+
+* Added the support for custom `bslib` bootstrap themes in `teal::init` apps, please read more in the new `teal-bs-themes` vignette.
 
 ### Enhancements